feat(api): Adiciona métodos GET para o recurso de denúncias

+++ backend/vozes-api/src/main/java/br/org/recode/vozes/model/Anexo.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class Anexo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idAnexo;
+    private Long idAnexo;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
